project refactor

+++ src/Helper.sol
@@ -2,27 +2,27 @@
 pragma solidity ^0.8.19;
 
 contract Helper {
-    function add(uint a, uint b) external pure returns (uint c) {
+    function add(uint256 a, uint256 b) external pure returns (uint256 c) {
         c = a + b;
     }
 
-    function sub(uint a, uint b) external pure returns (uint c) {
+    function sub(uint256 a, uint256 b) external pure returns (uint256 c) {
         c = a - b;
     }
 
-    function mul(uint a, uint b) external pure returns (uint c) {
+    function mul(uint256 a, uint256 b) external pure returns (uint256 c) {
         c = a * b;
     }
 
-    function div(uint a, uint b) external pure returns (uint c) {
+    function div(uint256 a, uint256 b) external pure returns (uint256 c) {
         c = a / b;
     }
 
-    function mod(uint a, uint b) external pure returns (uint c) {
+    function mod(uint256 a, uint256 b) external pure returns (uint256 c) {
         c = a % b;
     }
 
-    function exp(uint a, uint b) external pure returns (uint c) {
+    function exp(uint256 a, uint256 b) external pure returns (uint256 c) {
         c = a ** b;
     }
 }
